requirements.txt
dbt-snowflake==1.7.1

